requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Mail==0.9.1
email-validator==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.1.1
pytest==7.4.3
pytest-flask==1.3.0
Pillow==10.1.0